feat(compiler): double 全量干净实现完成 + 64位精度验证
- VM: 字节偏移架构（LoadLocal/StoreLocal/Call）、param_sizes 参数传递
- VM: 新增 LoadLocalD/StoreLocalD/LoadGlobalD/StoreGlobalD/LoadMemD/StoreMemD/SplitD
- BytecodeGen: 真正的 *D 指令生成（PushConstD/AddD/SubD/MulD/DivD/CastI2D/CastF2D）
- TypeChecker: FloatLiteral→Double 直接改 ty，VarDecl/Assign/InitList 补充隐式转换
- printf: %f 统一读取 f64，float 参数自动 CastF2D 提升
- sizeof(double) = 8
- 新增高精度测试 test_e2e_double_precision_64bit
- 216/216 测试通过，0 clippy 警告

+++ native/src/capi/mod.rs
@@ -833,7 +833,7 @@ pub unsafe extern "C" fn cide_variable_get(
     let is_arr = matches!(v.ty.kind, crate::compiler::ast::TypeKind::Array);
     if !is_array.is_null() { *is_array = if is_arr { 1 } else { 0 }; }
     if !array_size.is_null() { *array_size = if is_arr { v.ty.array_size } else { 0 }; }
-    if !value.is_null() { *value = v.value; }
+    if !value.is_null() { *value = v.value as c_int; }
 }
 
 #[no_mangle]
